perf(encoder): keep the sinusoidal PE in the activation dtype (3.5x faster float16 inference)
## What Changed
- `mlx_qwen3_asr/encoder.py`: `x = x + pe_full.astype(x.dtype)` in
  `AudioEncoder._encode_single`.
- `tests/test_model.py`: `TestEncoderDtype` casts a tiny encoder to float16 and
  asserts the output stays float16.

## Root Intent
`SinusoidalPositionEmbedding._pe` is built in float32 and, being an underscore
attribute, is outside `parameters()`, so `load_model()`'s dtype cast never
reached it. Adding it to float16 activations promoted the encoder output to
float32, and because those features are injected into the text embeddings the
whole decoder ran in float32 as well. The float16 activation clamp in
`AudioEncoderLayer` was unreachable.

Measured on Apple Silicon with `Qwen/Qwen3-ASR-0.6B`, dtype float16:

| | float32 PE (before) | float16 PE (after) |
|---|---|---|
| 15 s clip, median of 5 | 1.007 s | 0.333 s |
| LibriSpeech test-clean, 100 speaker-balanced samples, WER | 2.33% | 2.33% |
| CER | 0.59% | 0.59% |
| mean latency per sample | 1.532 s | 0.440 s |
| RTF | 0.171 | 0.049 |

All 100 hypotheses are byte-identical between the two runs. The upstream
PyTorch model registers the table as a buffer, so `.to(torch.float16)` casts
it there; this change restores parity with that behaviour.

## Seed Prompt
Any non-parameter buffer added to activations must be cast to the activation
dtype at the use site (or at load). Check with a probe that prints the dtype
of the encoder output and of the injected embeddings for a float16 model; both
must be float16. Gate dtype changes on `scripts/eval_librispeech.py --samples
100 --sampling speaker_round_robin` before and after, comparing WER and the
per-sample hypotheses in the JSON output, not just the aggregate.

## Files Changed
- mlx_qwen3_asr/encoder.py (+4/-1)
- tests/test_model.py (+20)

+++ mlx_qwen3_asr/encoder.py
@@ -414,7 +414,10 @@ class AudioEncoder(nn.Module):
         for ct in chunk_token_lens:
             pe_parts.append(pe[:ct])
         pe_full = mx.concatenate(pe_parts, axis=0)  # (total_tokens, d_model)
-        x = x + pe_full
+        # The PE table is float32 and outside parameters(), so load_model's dtype
+        # cast never touches it. Adding it unconverted would promote every later
+        # encoder (and decoder) activation to float32, tripling latency.
+        x = x + pe_full.astype(x.dtype)
 
         # --- Windowed attention ---
         total_tokens = x.shape[0]
